requirements.txt
-i https://pypi.org/simple
python-dotenv==0.21.0
slack-bolt==1.16.1
slack-sdk==3.19.5 ; python_full_version >= '3.6.0'
